+++ plugins/hh-geminio/src/main/kotlin/ru/hh/plugins/geminio/config/editor/GeminioPluginSettingsSearchableConfigurable.kt
@@ -42,11 +42,10 @@ class GeminioPluginSettingsSearchableConfigurable(
         return panel {
             row("Config File:") {
                 textFieldWithBrowseButton(
-                    browseDialogTitle = "Config File Path",
                     project = project,
                     fileChooserDescriptor = FileChooserDescriptorFactory.createSingleFileDescriptor(
                         PluginsConstants.YAML_FILES_FILTER_EXTENSION
-                    ),
+                    ).withTitle("Config File Path"),
                     fileChosen = { chosenFile ->
                         chosenFile.presentableUrl.also {
                             applyConfigurationFromFile(it)