+++ shared/feature/geminio-sdk/src/main/kotlin/ru/hh/plugins/geminio/sdk/recipe/models/optional/TemplateFormFactor.kt
@@ -11,7 +11,8 @@ enum class TemplateFormFactor(
     MOBILE("mobile"),
     WEAR("wear"),
     TV("tv"),
-    AUTOMOTIVE("automotive"),
+    CAR("car"),
+    XR("xr"),
     GENERIC("generic");
 
     companion object {